Move Path class to util module
